+++ Packages/Syntax/Sources/Syntax/Syntaxes/PHP/Queries/injections.scm
@@ -1,0 +1,13 @@
+;
+; Injections.scm
+; for PHP
+;
+;  CotEditor
+;  https://coteditor.com
+;
+; © 2026 1024jp
+;
+
+((text) @injection.content
+ (#set! injection.language "html")
+ (#set! injection.combined))