+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 ### Improvements
 
 - Improve the syntax highlighting algorithm.
-- Migrate the built-in CSS, HTML, JavaScript, Python, Ruby, and Swift syntaxes to use tree-sitter for syntax highlighting.
+- Migrate the built-in CSS, HTML, JavaScript, PHP, Python, Ruby, and Swift syntaxes to use tree-sitter for syntax highlighting.
 - Update built-in syntaxes.
 - Improve the performance of general text editing.
 - [trivial] Adjust the About CotEditor window.